+++ src/main/kotlin/id/walt/model/dif/InputDescriptorField.kt
@@ -1,8 +1,11 @@
 package id.walt.model.dif
 
+import com.beust.klaxon.JsonObject
+import id.walt.model.oidc.JsonObjectField
+
 data class InputDescriptorField(
     val path: List<String>,
     val id: String? = null,
     val purpose: String? = null,
-    val filter: Map<String, Any>? = null
+    @JsonObjectField val filter: Map<String, Any?>? = null
 )